move user dict builder to own file
